Various stuff
pdf2json/pdfcollect.ml knows how to collect pdf elements such as text,
lines and boxes, but its implementation is a bit ugly and should be
extended (eg. to compute the outer box of the collected text elements).

statemachine.ml contains a cleaner draft of the implementation of the
inner state machine used for collecting the pdf elements.